started to deal with #2
