+++ fmriprep/__init__.py
@@ -1,0 +1,6 @@
+
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+from .workflows.base import fmri_preprocess
