+++ fmriprep/viz/variables_reports.py
